README.md
We've established a foundational process for IT support wherein, upon a customer's call to the IT support number, we leverage call recording. Additionally, we've integrated a Lambda function for holiday checks, which seamlessly connects with a DynamoDB table. During holidays, we offer a callback service, and in other instances, we provide a self-service menu integrated with Lex boxes featuring four intents. This serves as a basic example of the capabilities of Amazon Connect, ranging from setting up queues and agent profiles to routing profiles and Quick Connect. We can further leverage the AWS ecosystem, incorporating services like CloudWatch, QuickSight, and Lambda, among others.